requirements.txt
aniso8601==8.0.0
APScheduler==3.6.3
Babel==2.8.0
base58==2.0.1
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
coinpayments-py==0.1.0
cytoolz==0.10.1
dnspython==2.0.0
ecdsa==0.15
email-validator==1.1.1
enum34==1.1.6
eth-abi==2.1.1
eth-hash==0.2.0
eth-typing==2.2.2
eth-utils==1.9.5
Flask==1.1.2
Flask-Admin==1.5.6
Flask-BabelEx==0.9.4
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Security==3.0.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
parsimonious==0.8.1
passlib==1.7.2
pycparser==2.20
pycryptodome==3.9.8
pyTelegramBotAPI==3.7.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.19
toolz==0.10.0
tronpy==0.1.4
tzlocal==2.1
urllib3==1.25.10
Werkzeug==1.0.1
WTForms==2.3.3
